API: Extracted JsonApiController module.
Handles the routine mechanics of rendering the right responses.
