+++ app/controllers/forem/api/v1/forums_controller.rb
@@ -4,15 +4,7 @@ module Forem
   module Api
     module V1
       class ForumsController < Forem::ForumsController
-        rescue_from CanCan::AccessDenied, with: :forbidden
-
-        before_action { request.format = :json }
-
-        private
-
-        def forbidden
-          render nothing: true, status: :forbidden
-        end
+        include JsonApiController
       end
     end
   end
