+++ app/controllers/forem/api/v1/topics_controller.rb
@@ -4,46 +4,20 @@ module Forem
   module Api
     module V1
       class TopicsController < Forem::TopicsController
-        before_action { request.format = :json }
+        include JsonApiController
 
-        before_action :client_generated_ids_are_unsupported,
-          only: [:create, :update]
-
-        rescue_from ActionController::ParameterMissing, with: :bad_request
-        rescue_from CanCan::AccessDenied, with: :forbidden
-
-        protected
-
-        def authenticate_forem_user
-          render nothing: true, status: :forbidden if !forem_user
-        end
+        private
 
         def topic_params
-          params.require(:data).require(:attributes).permit(:subject, :posts_attributes => [[:text]])
+          required_params.permit(:subject)
         end
 
-        def create_successful
-          render 'show', status: :created, location: api_forum_topic_url(@forum, @topic)
-        end
-
-        def create_unsuccessful
-          render 'member_errors', status: :bad_request
+        def resource_url
+          api_forum_topic_url(@forum, @topic)
         end
 
         def topic_not_found
           render nothing: true, status: :not_found
-        end
-
-        def bad_request
-          render nothing: true, status: :bad_request
-        end
-
-        def forbidden
-          render nothing: true, status: :forbidden
-        end
-
-        def client_generated_ids_are_unsupported
-          render nothing: true, status: :forbidden if params[:data][:id]
         end
       end
     end
